+++ common/constants/collectionNames.js
@@ -1,0 +1,5 @@
+const collectionNames = {
+    accessories: 'accessories'
+};
+
+export default collectionNames;