+++ common/constants/errorCodes.js
@@ -1,0 +1,9 @@
+const errorCodes = {
+    'created': 201,
+    'deleted': 202,
+    'badRequest': 400,
+    'notFound': 404,
+    'duplicatedField': 11000,
+};
+
+export default errorCodes;