+++ common/schemas/index.js
@@ -1,0 +1,6 @@
+import CollectionNames from '../constants/collectionNames.js';
+import accessoriesSchema from './accessoriesSchema.js';
+
+export default {
+    [CollectionNames.accessories]: accessoriesSchema
+}