+++ server/src/index.js
@@ -7,7 +7,6 @@ import cookieParser from 'cookie-parser';
 import mongoose from 'mongoose';
 import Startup from './startup/index.js';
 import Routes from './routes/index.js';
-import authMiddleware from './middleware/auth.middleware.js';
 
 const app = express();
 
@@ -24,12 +23,12 @@ app.use(cookieParser());
 // routes
 app.use('/api/status', (req, res) => res.send('OK'));
 app.use('/api/auth', Routes.AuthRoutes);
-app.use('/api/accessories', authMiddleware, Routes.AccessorieRoutes);
-app.use('/api/assets', authMiddleware, Routes.AssetRoutes);
-app.use('/api/licenses', authMiddleware, Routes.LicenseRoutes);
-app.use('/api/location', authMiddleware, Routes.LocationRoutes);
-app.use('/api/company', authMiddleware, Routes.CompanyRoutes);
-app.use('/api/department', authMiddleware, Routes.DepartmentRoutes);
+app.use('/api/accessories', Routes.AccessorieRoutes);
+app.use('/api/assets', Routes.AssetRoutes);
+app.use('/api/licenses', Routes.LicenseRoutes);
+app.use('/api/location', Routes.LocationRoutes);
+app.use('/api/company', Routes.CompanyRoutes);
+app.use('/api/department', Routes.DepartmentRoutes);
 
 // connections
 mongoose