+++ server/src/lib/collections/index.js
@@ -1,0 +1,4 @@
+import Collections from 'asset-management-common/schemas/index.js';
+import CollectionNames from 'asset-management-common/constants/collectionNames.js';
+
+export const Accessories = Collections[CollectionNames.accessories];